Add files via upload
Rewrite the superclass Shape and its subclasses Circle, Rectangle and Square, as shown in the class
diagram.
In this exercise, Shape shall be defined as an abstract class, which contains:

Two protected instance variables color(String) and filled(boolean). The protected variables can be accessed
by its subclasses and classes in the same package. They are denoted with a '#' sign in the class diagram.
Getter and setter for all the instance variables, and toString().
Two abstract methods getArea() and getPerimeter() (shown in italics in the class diagram).
The subclasses Circle and Rectangle shall override the abstract methods getArea() and getPerimeter() and
provide the proper implementation. They also override the toString().

Write a test class to test these statements involving polymorphism and explain the outputs. Some statements
may trigger compilation errors. Explain the errors, if any.